Fixed mobile menu still closing if clicked inside
Fixed a bug where the mobile menu sidebar still closes even if the user
clicks inside the sidebar.

The expected behaviour was that the sidebar only closes when the user
clicks outside the sidebar or clicks the x button.

It was fixed by adding a click event listener on the sidebar's backdrop
and making the sidebar close when the backdrop was clicked.